+++ all-is-cubes-port/src/mv/palette.rs
@@ -1,24 +1,65 @@
 use std::sync::Arc;
 
+use itertools::{EitherOrBoth, Itertools};
+
+use all_is_cubes::arcstr;
 use all_is_cubes::block::{self, Block};
-use all_is_cubes::math::Rgba;
+use all_is_cubes::math::{Rgb, Rgba, ZeroOne};
 
 // -------------------------------------------------------------------------------------------------
 
 #[cfg(feature = "import")]
-pub(crate) fn dot_vox_palette_to_blocks(palette: &[dot_vox::Color]) -> Arc<[Block]> {
+pub(crate) fn dot_vox_palette_to_blocks(
+    palette: &[dot_vox::Color],
+    materials: &[dot_vox::Material],
+) -> Arc<[Block]> {
     palette
         .iter()
+        .zip_longest(materials)
         .enumerate()
-        .map(|(index, &dot_vox::Color { r, g, b, a })| {
+        .map(|(index, palette_color_and_or_material)| {
+            let display_name = arcstr::format!("{index}");
+            let (reflectance, emission) = match palette_color_and_or_material {
+                EitherOrBoth::Both(&color, material) => {
+                    let palette_color = color_in(color);
+                    match material.material_type() {
+                        Some("_blend" | "_glass") => (
+                            palette_color.with_alpha(
+                                ZeroOne::<f32>::try_from(material.opacity().unwrap_or(1.0))
+                                    .unwrap_or(ZeroOne::ONE)
+                                    .complement(),
+                            ),
+                            Rgb::ZERO,
+                        ),
+                        Some("_emit") => {
+                            // Empirically, radiant_flux scales the emission by powers of 10.
+                            let emission_scale = material.emission().unwrap_or(0.0)
+                                * 10.0f32.powf(material.radiant_flux().unwrap_or(0.0));
+                            (Rgba::BLACK, palette_color * emission_scale)
+                        }
+                        None | Some("_diffuse") => (palette_color.with_alpha_one(), Rgb::ZERO),
+                        _ => {
+                            log::warn!("unknown/unsupported .vox material type {material:?}");
+                            (palette_color.with_alpha_one(), Rgb::ZERO)
+                        }
+                    }
+                }
+                EitherOrBoth::Left(&color) => (color_in(color).with_alpha_one(), Rgb::ZERO),
+                EitherOrBoth::Right(_material_only) => {
+                    log::warn!("material but no palette at index {index}");
+                    (Rgba::WHITE, Rgb::ZERO)
+                }
+            };
             Block::builder()
-                .display_name(index.to_string())
-                .color(Rgba::from_srgb8([r, g, b, a]))
+                .display_name(display_name)
+                .color(reflectance)
+                .light_emission(emission)
                 .build()
         })
         .collect()
 }
 
+// TODO: export materials (this will require support from dot_vox)
 #[cfg(feature = "export")]
 pub(crate) fn block_to_dot_vox_palette_entry(
     evaluated: &block::EvaluatedBlock,
@@ -30,4 +71,11 @@ pub(crate) fn block_to_dot_vox_palette_entry(
         let [r, g, b, a] = evaluated.color().to_srgb8();
         Some(dot_vox::Color { r, g, b, a })
     }
+}
+
+#[cfg(feature = "import")]
+fn color_in(dot_vox::Color { r, g, b, a: _ }: dot_vox::Color) -> Rgb {
+    // Note: Even though the `dot_vox::Color` field carries an alpha, MagicaVoxel itself seems to
+    // always ignore it, and use only transparency specified by materials. Therefore, we do too.
+    Rgb::from_srgb8([r, g, b])
 }